feat: complete KoLCharacter depth remediation
New types:
- Gender.kt: UNKNOWN/MALE/FEMALE enum with modifier value and API string parsing
- AscensionPath.kt: major paths with canEat/canDrink/canChew/hasPathResources flags;
  fromApiString() mapping; isAvatar computed property

CharacterState — comprehensive expansion covering every desktop field category:

  Identity: gender, title (in addition to existing fields)

  HP/MP: baseMaxHp, baseMaxMp — HP/MP cap before equipment/effect buffs

  Stats: muscSubpoints/mystSubpoints/moxieSubpoints promoted from Int to Long
    (desktop uses long for subpoints; avoids overflow at high-level characters);
    muscleTNP, mystTNP, moxieTNP, mainStatTNP computed (till-next-point in subpoints)

  Currency: storageMeat (Hagnk's), closetMeat, sessionMeat (runtime delta)

  Turns: currentRun (turns this ascension, separate from all-time turnsPlayed),
    globalDaycount (KoL calendar day), rolloverTimestamp (Unix epoch)

  Path-aware consumption: canEat, canDrink, canChew, isFallingDown, isOverdrunk
    computed from AscensionPath flags + consumable counts

  Rollover: secondsUntilRollover, isNearRollover

  Class-specific resources (all safe-default 0 for non-applicable classes):
    fury (Seal Clubber), soulsauce (Sauceror), discoMomentum (Disco Bandit),
    audience (Accordion Thief), absorbs (Grey Goo path),
    thunder/rain/lightning + currentPP/maximumPP (Plumber path),
    youRobotEnergy/Scraps (You, Robot), wildfireWater (Wildfire), minstrelLevel (Boris)

  Campground: telescopeUpgrades (0-7), hasBookshelf

  Social: hasStore, hasDisplayCase, hasClan, hippyStoneBroken, kingLiberated

  Familiar: enthronedFamiliarId/Name (Crown of Thrones),
    bjornedFamiliarId/Name (Buddy Bjorn), arenaWins
    hasEnthroned, hasBjorned computed

  Misc: mindControlLevel, radSickness, stillsAvailable, arenaWins
    isUnarmed computed from equipment map
    furyLimit, audienceLimit class resource ceilings
    canUseStore, canUseDisplayCase access flags

CharacterApiResponse — full field coverage matching CharacterState:
  basehpmax, basempmax, currentrun, rollover, storagemeat,
  fury, soulsauce, thunder, rain, lightning, pp, ppmax,
  robonenergy, robonscraps, wildfirewater, minstrel,
  hasstore, hasdisplaycase, hasclan, hippystone,
  telescopelevel, kingliberated, gender, title, mcd,
  radsickness, stills, arenawins,
  enthroned/enthronedname, bjorned/bjornedname

KoLCharacter:
  - updateFromApiResponse populates every new field with safe fallbacks
  - updateMeat() extended with optional storageMeat param
  - addSessionMeat(delta) for runtime meat tracking
  - updateClassResource() for partial class resource patches
  - updatePlumberResources() for thunder/rain/lightning/PP
  - setCampground() for telescope + bookshelf state
  - setMindControlLevel(), setCurrentRun() partial updaters

UI fix: CharacterScreen.StatRow subpoints param Int→Long to match new type

+++ shared/src/commonMain/kotlin/net/sourceforge/kolmafia/ui/character/CharacterScreen.kt
@@ -155,7 +155,7 @@ private fun StatBar(label: String, current: Int, max: Int) {
 }
 
 @Composable
-private fun StatRow(label: String, base: Int, subpoints: Int) {
+private fun StatRow(label: String, base: Int, subpoints: Long) {
     Row(
         modifier = Modifier.fillMaxWidth().padding(vertical = 2.dp),
         horizontalArrangement = Arrangement.SpaceBetween
